delete branch

+++ Backup Service/src/server/ControlChannelThread.java
@@ -280,7 +280,13 @@ public class ControlChannelThread extends ChannelThread{
 	    File file = new File(Values.directory_to_backup_files+fileSeparator+ message.getFileID()); 
 	    
 	    if(file.isDirectory() && file.exists()){  
-	        file.delete();
+	        File[] chunks = file.listFiles();
+	        for(File f : chunks) {
+	            f.delete();
+	        }
+	        if(file.list().length == 0) {
+	            file.delete();
+	        }
 	    } else {
 	        System.out.println("RECEIVED DELETE MSG FOR FILE "+message.getFileID()+" THAT IS NOT BACKED UP IN THIS PEER");
 	    }
